Kibana Error Boundary - Remove error collection from EBT
