+++ src/platform/packages/shared/shared-ux/error_boundary/index.ts
@@ -11,6 +11,3 @@ export { KibanaErrorBoundaryProvider } from './src/services/error_boundary_provi
 export { KibanaErrorBoundary } from './src/ui/error_boundary';
 export { KibanaSectionErrorBoundary } from './src/ui/section_error_boundary';
 export { ThrowIfError } from './src/ui/throw_if_error';
-
-export { REACT_FATAL_ERROR_EVENT_TYPE, reactFatalErrorSchema } from './lib/telemetry_events';
-export type { ReactFatalError } from './lib/telemetry_events';
