+++ src/platform/packages/shared/shared-ux/error_boundary/mocks/src/jest.ts
@@ -11,11 +11,8 @@ import { KibanaErrorService } from '../../src/services/error_service';
 import type { KibanaErrorBoundaryServices } from '../../types';
 
 export const getServicesMock = (): KibanaErrorBoundaryServices => {
-  const mockDeps = {
-    analytics: { reportEvent: jest.fn() },
-  };
   return {
     onClickRefresh: jest.fn().mockResolvedValue(undefined),
-    errorService: new KibanaErrorService(mockDeps),
+    errorService: new KibanaErrorService(),
   };
 };
