+++ src/platform/packages/shared/shared-ux/error_boundary/mocks/src/storybook.ts
@@ -10,7 +10,6 @@
 import { AbstractStorybookMock } from '@kbn/shared-ux-storybook-mock';
 import { action } from '@storybook/addon-actions';
 import { KibanaErrorService } from '../../src/services/error_service';
-import { createAnalyticsMock } from './analytics_mock';
 import type { KibanaErrorBoundaryServices } from '../../types';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
@@ -28,13 +27,11 @@ export class KibanaErrorBoundaryStorybookMock extends AbstractStorybookMock<
 
   getServices(params: Params = {}): KibanaErrorBoundaryServices {
     const onClickRefresh = action('Reload window');
-    const mock = createAnalyticsMock();
-    const analytics = mock.analytics;
 
     return {
       ...params,
       onClickRefresh,
-      errorService: new KibanaErrorService({ analytics }),
+      errorService: new KibanaErrorService(),
     };
   }
