+++ src/platform/packages/shared/shared-ux/error_boundary/moon.yml
@@ -19,7 +19,6 @@ project:
 dependsOn:
   - '@kbn/shared-ux-storybook-mock'
   - '@kbn/i18n'
-  - '@kbn/core-analytics-browser-mocks'
 tags:
   - shared-common
   - package
