+++ src/platform/packages/shared/shared-ux/error_boundary/src/services/error_boundary_provider.test.tsx
@@ -10,57 +10,35 @@
 import { render } from '@testing-library/react';
 import React from 'react';
 
-import { analyticsServiceMock } from '@kbn/core-analytics-browser-mocks';
-import type { KibanaErrorBoundaryProviderDeps } from '../../types';
 import { KibanaErrorBoundary, KibanaErrorBoundaryProvider } from '../..';
 import { BadComponent } from '../../mocks';
-import { TRANSIENT_NAVIGATION_WINDOW_MS } from './error_service';
+import { errorMessageStrings as strings } from '../ui/message_strings';
 import userEvent from '@testing-library/user-event';
 
 describe('<KibanaErrorBoundaryProvider>', () => {
-  let analytics: KibanaErrorBoundaryProviderDeps['analytics'];
   beforeEach(() => {
     jest.spyOn(console, 'error').mockImplementation(() => {});
-    analytics = analyticsServiceMock.createAnalyticsServiceStart();
   });
 
   it('creates a context of services for KibanaErrorBoundary', async () => {
-    const reportEventSpy = jest.spyOn(analytics!, 'reportEvent');
-
-    const { findByTestId, unmount } = render(
-      <KibanaErrorBoundaryProvider analytics={analytics}>
+    const { findByTestId, findByText } = render(
+      <KibanaErrorBoundaryProvider>
         <KibanaErrorBoundary>
           <BadComponent />
         </KibanaErrorBoundary>
       </KibanaErrorBoundaryProvider>
     );
     await userEvent.click(await findByTestId('clickForErrorBtn'));
-    unmount(); // Unmount to commit/report the error
 
-    // Wait for the error to be reported/committed
-    await new Promise((resolve) => setTimeout(resolve, 1.5 * TRANSIENT_NAVIGATION_WINDOW_MS));
-
-    expect(reportEventSpy).toBeCalledWith('fatal-error-react', {
-      component_name: 'BadComponent',
-      component_stack: expect.any(String),
-      error_message: 'Error: This is an error to show the test user!',
-      error_stack: expect.any(String),
-      component_render_min_duration_ms: expect.any(Number),
-      has_transient_navigation: expect.any(Boolean),
-    });
+    expect(await findByText(strings.page.callout.fatal.title())).toBeVisible();
   });
 
   it('uses higher-level context if available', async () => {
-    const reportEventParentSpy = jest.spyOn(analytics!, 'reportEvent');
-
-    const analyticsChild = analyticsServiceMock.createAnalyticsServiceStart();
-    const reportEventChildSpy = jest.spyOn(analyticsChild, 'reportEvent');
-
-    const { findByTestId, unmount } = render(
-      <KibanaErrorBoundaryProvider analytics={analytics}>
+    const { findByTestId, findByText } = render(
+      <KibanaErrorBoundaryProvider>
         <KibanaErrorBoundary>
           Hello world
-          <KibanaErrorBoundaryProvider analytics={analyticsChild}>
+          <KibanaErrorBoundaryProvider>
             <KibanaErrorBoundary>
               <BadComponent />
             </KibanaErrorBoundary>
@@ -70,26 +48,6 @@ describe('<KibanaErrorBoundaryProvider>', () => {
     );
     await userEvent.click(await findByTestId('clickForErrorBtn'));
 
-    // Wait for nav to settle
-    await new Promise((resolve) => setTimeout(resolve, TRANSIENT_NAVIGATION_WINDOW_MS));
-
-    unmount(); // Unmount to commit/report the error
-
-    // Wait for the error to be reported/committed
-    await new Promise((resolve) => setTimeout(resolve, 500));
-
-    expect(reportEventParentSpy).not.toBeCalled();
-    expect(reportEventChildSpy).toBeCalledWith('fatal-error-react', {
-      component_name: 'BadComponent',
-      component_stack: expect.any(String),
-      error_message: 'Error: This is an error to show the test user!',
-      error_stack: expect.any(String),
-      has_transient_navigation: expect.any(Boolean),
-      component_render_min_duration_ms: expect.any(Number),
-    });
-    expect(
-      (reportEventChildSpy.mock.calls[0][1] as Record<string, unknown>)
-        .component_render_min_duration_ms
-    ).toBeGreaterThanOrEqual(250);
+    expect(await findByText(strings.page.callout.fatal.title())).toBeVisible();
   });
 });
