+++ src/platform/plugins/shared/share/moon.yml
@@ -26,8 +26,6 @@ dependsOn:
   - '@kbn/core-custom-branding-browser'
   - '@kbn/core-saved-objects-utils-server'
   - '@kbn/react-kibana-context-theme'
-  - '@kbn/core-analytics-browser'
-  - '@kbn/shared-ux-error-boundary'
   - '@kbn/core-chrome-browser'
   - '@kbn/shared-ux-prompt-not-found'
   - '@kbn/react-kibana-mount'
