+++ src/platform/plugins/shared/share/public/plugin.ts
@@ -24,7 +24,6 @@ import type { BrowserShortUrlClient } from './url_service/short_urls/short_url_c
 import type { AnonymousAccessServiceContract } from '../common';
 import { LegacyShortUrlLocatorDefinition } from '../common/url_service/locators/legacy_short_url_locator';
 import { ShortUrlRedirectLocatorDefinition } from '../common/url_service/locators/short_url_redirect_locator';
-import { registrations } from './lib/registrations';
 import type { BrowserUrlService } from './types';
 
 /** @public */
@@ -90,7 +89,7 @@ export class SharePlugin
   constructor(private readonly initializerContext: PluginInitializerContext) {}
 
   public setup(core: CoreSetup): SharePublicSetup {
-    const { analytics, http } = core;
+    const { http } = core;
     const { basePath } = http;
 
     this.url = new UrlService<BrowserShortUrlClientFactoryCreateParams, BrowserShortUrlClient>({
@@ -127,8 +126,6 @@ export class SharePlugin
     });
     this.redirectManager.registerLocatorRedirectApp(core);
     this.redirectManager.registerLegacyShortUrlRedirectApp(core);
-
-    registrations.setup({ analytics });
 
     return {
       ...this.shareRegistry.setup(),
