+++ src/platform/plugins/shared/share/tsconfig.json
@@ -19,8 +19,6 @@
     "@kbn/core-custom-branding-browser",
     "@kbn/core-saved-objects-utils-server",
     "@kbn/react-kibana-context-theme",
-    "@kbn/core-analytics-browser",
-    "@kbn/shared-ux-error-boundary",
     "@kbn/core-chrome-browser",
     "@kbn/shared-ux-prompt-not-found",
     "@kbn/react-kibana-mount",
